rename dir to org.dark0ghost.configs & org.dark0ghost.annotations
