+++ src/main/java/igorlink/command/FilterSubCommand.java
@@ -5,8 +5,8 @@ import igorlink.service.Utils;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import static org.dark0ghost.annotations.configs.Text.RU.FilterSubCommandText.FILTER_OFF_TEXT;
-import static org.dark0ghost.annotations.configs.Text.RU.FilterSubCommandText.FILTER_ON_TEXT;
+import static org.dark0ghost.configs.Text.RU.FilterSubCommandText.FILTER_OFF_TEXT;
+import static org.dark0ghost.configs.Text.RU.FilterSubCommandText.FILTER_ON_TEXT;
 
 
 public class FilterSubCommand {
