+++ src/main/java/igorlink/donationexecutor/executionsstaff/Donation.java
@@ -5,7 +5,7 @@ import kotlin.Suppress;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 
-import static org.dark0ghost.annotations.configs.Text.RU.DonationText.DEFAULT_NAME;
+import static org.dark0ghost.configs.Text.RU.DonationText.DEFAULT_NAME;
 
 public class Donation {
     private CommandSender sender;
