+++ src/main/kotlin/org/dark0ghost/annotations/ExperimentalApi.kt
@@ -1,4 +1,4 @@
-package org.dark0ghost.annotations.annotations
+package org.dark0ghost.annotations
 
 @RequiresOptIn(message = "This API is experimental. It may be changed in the future without notice.")
 @Retention(AnnotationRetention.BINARY)
