+++ src/main/kotlin/org/dark0ghost/annotations/MagicConst.kt
@@ -1,4 +1,4 @@
-package org.dark0ghost.annotations.annotations
+package org.dark0ghost.annotations
 
 @Target(AnnotationTarget.FIELD)
 annotation class MagicConst
