+++ src/main/kotlin/org/dark0ghost/configs/Text.kt
@@ -1,4 +1,4 @@
-package org.dark0ghost.annotations.configs
+package org.dark0ghost.configs
 
 object Text {
     object RU {
